move the ref up and add tooltip v1 context and update snapshots
